+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 
 If you like this project and find it useful, please consider giving it a star on GitHub at https://github.com/Luligu/matterbridge and sponsoring it.
 
+## [1.3.13] - 2024-07-11
+
+### Added
+
+### Changed
+
+- [frontend]: The Logs window in the Home page has the same filter as the Logs page.
+- [matterbridge]: The plugins debug is now indipendent from matterbridge debug and matter.js log level. It can be set from the plugin config.
+
+### Fixed
+
+- [frontend]: Fix Home page for mobile (the page doesn't "jump" anymore with touchscreens).
+- [matterbridge]: Fixed npm ignore for exports.
+- [matterbridge]: Fixed load plugin when the don't have author and description.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
 ## [1.3.12] - 2024-07-10
 
 ### Added